moved DeleteListIdType to IdColumnType
