+++ impexp-config/src/main/resources/org/citydb/config/i18n/language_de.properties
@@ -145,6 +145,8 @@ filter.label.lod.depth=Suchtiefe
 filter.label.id=Identifier
 filter.label.name=gml:name
 filter.label.lineage=citydb:lineage
+filter.idList.id.resource=Objekt-ID
+filter.idList.id.database=Datenbank-ID
 filter.label.counter.count=Anzahl
 filter.label.counter.startIndex=Startindex
 filter.label.sql.increase=SQL-Eingabefeld vergrößern
@@ -210,8 +212,6 @@ delete.dialog.title.delete=Lösche Stadtobjekte...
 delete.dialog.title.terminate=Terminiere Stadtobjekte...
 delete.dialog.title.preview=Erzeuge Löschvorschau...
 delete.dialog.title.cleanupGlobalAppearances=Lösche globale Appearances...
-delete.list.id.resource=Objekt-ID
-delete.list.id.database=Datenbank-ID
 
 #VisExportPanel
 visExport.label.automatic=Feste Seitenlänge
