+++ impexp-config/src/main/resources/org/citydb/config/i18n/language_en.properties
@@ -145,6 +145,8 @@ filter.label.lod.depth=Search depth
 filter.label.id=Identifier
 filter.label.name=gml:name
 filter.label.lineage=citydb:lineage
+filter.idList.id.resource=Object ID
+filter.idList.id.database=Database ID
 filter.label.counter.count=Count
 filter.label.counter.startIndex=Start index
 filter.label.sql.increase=Increase SQL text field
@@ -210,8 +212,6 @@ delete.dialog.title.delete=Deleting city objects...
 delete.dialog.title.terminate=Terminating city objects...
 delete.dialog.title.preview=Creating delete preview...
 delete.dialog.title.cleanupGlobalAppearances=Cleaning up global appearances...
-delete.list.id.resource=Object ID
-delete.list.id.database=Database ID
 
 #VisExportPanel
 visExport.label.automatic=Fixed side length
