+++ impexp-core/src/main/java/org/citydb/core/operation/deleter/database/DeleteListImporter.java
@@ -28,7 +28,7 @@
 
 package org.citydb.core.operation.deleter.database;
 
-import org.citydb.config.project.deleter.DeleteListIdType;
+import org.citydb.config.project.common.IdColumnType;
 import org.citydb.core.database.schema.mapping.MappingConstants;
 import org.citydb.core.operation.common.cache.CacheTable;
 import org.citydb.core.operation.common.csv.IdListException;
@@ -48,9 +48,9 @@ public class DeleteListImporter {
         this.maxBatchSize = maxBatchSize;
     }
 
-    public void doImport(IdListParser parser, DeleteListIdType idType) throws IdListException, SQLException {
+    public void doImport(IdListParser parser, IdColumnType idColumnType) throws IdListException, SQLException {
         String sql = "insert into " + cacheTable.getTableName() + " " +
-                (idType == DeleteListIdType.DATABASE_ID ?
+                (idColumnType == IdColumnType.DATABASE_ID ?
                         "(" + MappingConstants.ID + ") values (?)" :
                         "(" + MappingConstants.GMLID + ") values (?)");
 
@@ -60,7 +60,7 @@ public class DeleteListImporter {
                 String id = parser.nextId();
 
                 if (id != null) {
-                    if (idType == DeleteListIdType.DATABASE_ID) {
+                    if (idColumnType == IdColumnType.DATABASE_ID) {
                         try {
                             ps.setLong(1, Long.parseLong(id));
                         } catch (NumberFormatException e) {
